+++ MOP/src/MOP.cs
@@ -23,9 +23,9 @@ namespace MOP
     public class MOP : Mod
     {
         public override string ID => "MOP"; //Your mod ID (unique)
-        public override string Name => "Modern Optimization Plugin"; //You mod name
+        public override string Name => "Modern Optimization Plugin (Development)"; //You mod name
         public override string Author => "Athlon"; //Your Username
-        public override string Version => "2.1.2"; //Version
+        public override string Version => "2.1.3"; //Version
 
         // Set this to true if you will be load custom assets from Assets folder.
         // This will create subfolder in Assets folder for your mod.